Extract FlashSaleAuditLog into its own class
